int mult -- double mult

+++ Main.java
@@ -12,7 +12,7 @@ public class Main
             summe += 1;
         }
         
-        int mult = 1;
+        double mult = 1;
         for(int i = 1; i <= obereGrenze; ++i)
         {
             mult *= i;
